fix: revert to old selectors
